requirements.txt
# Energie-Monitor – Dependencies
PyQt6>=6.6
psycopg2-binary>=2.9
python-dotenv>=1.0
matplotlib>=3.8
